requirements.txt
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
firebase-admin==3.2.1
future==0.18.2
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
httplib2==0.15.0
idna==2.8
msgpack==0.6.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.7
